Unattended installation of puppet agent and session lifetime increase.
The bootstrapping process now applies various platform class to the
puppet master to be able to subsequently and automatically install the
puppet agent on alle nodes including registering the nodes on the puppet
master.
Additionally the session lifetime for the dashboard is increased
to a configurable number of seconds (@see Vagrantfile).